v3.0.0: recall glyph now works with runestones

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/stones/StoneOfBlink.java
@@ -53,7 +53,7 @@ public class StoneOfBlink extends Runestone {
 	
 	@Override
 	protected void activate(int cell) {
-		if (!ScrollOfTeleportation.teleportToLocation(curUser, cell)){
+		if (!ScrollOfTeleportation.teleportToLocation(curUser, cell) && !anonymous){
 			Dungeon.level.drop(this, cell).sprite.drop();
 		}
 	}
